Fix Integration tests against 'minified' sources
