+++ packages/firestore/test/integration/api/database.test.ts
@@ -20,13 +20,9 @@ import * as chaiAsPromised from 'chai-as-promised';
 import * as firestore from '@firebase/firestore-types';
 import { expect, use } from 'chai';
 
-import { SimpleDb } from '../../../src/local/simple_db';
-import { fail } from '../../../src/util/assert';
-import { Code } from '../../../src/util/error';
-import { query } from '../../util/api_helpers';
 import { Deferred } from '../../util/promise';
 import { EventsAccumulator } from '../util/events_accumulator';
-import firebase from '../util/firebase_export';
+import * as firebase from '../util/firebase_export';
 import {
   apiDescribe,
   withTestCollection,
@@ -34,17 +30,19 @@ import {
   withTestDbs,
   withTestDoc,
   withTestDocAndInitialData,
-  DEFAULT_SETTINGS,
-  withMockCredentialProviderTestDb
+  DEFAULT_SETTINGS
 } from '../util/helpers';
-import { User } from '../../../src/auth/user';
 
 // tslint:disable:no-floating-promises
 
 use(chaiAsPromised);
 
-const Timestamp = firebase.firestore!.Timestamp;
-const FieldValue = firebase.firestore!.FieldValue;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const Timestamp = (firebase as any).firestore.Timestamp;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const FieldPath = (firebase as any).firestore.FieldPath;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const FieldValue = (firebase as any).firestore.FieldValue;
 
 apiDescribe('Database', (persistence: boolean) => {
   it('can set a document', () => {
@@ -134,8 +132,10 @@ apiDescribe('Database', (persistence: boolean) => {
         .get({ source: 'cache' })
         .then(doc => expect(doc.exists).to.be.true);
       await readerRef.get({ source: 'cache' }).then(
-        () => fail('Expected cache miss'),
-        err => expect(err.code).to.be.equal(Code.UNAVAILABLE)
+        () => {
+          throw new Error('Expected cache miss');
+        },
+        err => expect(err.code).to.be.equal('unavailable')
       );
       await writerRef
         .get()
@@ -488,8 +488,6 @@ apiDescribe('Database', (persistence: boolean) => {
   });
 
   it('can update nested fields', () => {
-    const FieldPath = firebase.firestore!.FieldPath;
-
     return withTestDoc(persistence, doc => {
       const initialData = {
         desc: 'Description',
@@ -818,67 +816,46 @@ apiDescribe('Database', (persistence: boolean) => {
   // have security rules support or something?
   // eslint-disable-next-line no-restricted-properties
   describe.skip('Listens are rejected remotely:', () => {
-    const queryForRejection = query('foo');
-
     it('will reject listens', () => {
-      const deferred = new Deferred();
-      queryForRejection.onSnapshot(
-        () => {},
-        (err: Error) => {
-          expect(err.name).to.exist;
-          expect(err.message).to.exist;
-          deferred.resolve();
-        }
-      );
-      return deferred.promise;
+      return withTestCollection(persistence, {}, queryForRejection => {
+        const deferred = new Deferred<void>();
+        queryForRejection.onSnapshot(
+          () => {},
+          (err: Error) => {
+            expect(err.name).to.exist;
+            expect(err.message).to.exist;
+            deferred.resolve();
+          }
+        );
+        return deferred.promise;
+      });
     });
 
     it('will reject same listens twice in a row', () => {
-      const deferred = new Deferred();
-      queryForRejection.onSnapshot(
-        () => {},
-        (err: Error) => {
-          expect(err.name).to.exist;
-          expect(err.message).to.exist;
-          queryForRejection.onSnapshot(
-            () => {},
-            (err2: Error) => {
-              expect(err2.name).to.exist;
-              expect(err2.message).to.exist;
-              deferred.resolve();
-            }
-          );
-        }
-      );
-      return deferred.promise;
+      return withTestCollection(persistence, {}, queryForRejection => {
+        const deferred = new Deferred<void>();
+        queryForRejection.onSnapshot(
+          () => {},
+          (err: Error) => {
+            expect(err.name).to.exist;
+            expect(err.message).to.exist;
+            queryForRejection.onSnapshot(
+              () => {},
+              (err2: Error) => {
+                expect(err2.name).to.exist;
+                expect(err2.message).to.exist;
+                deferred.resolve();
+              }
+            );
+          }
+        );
+        return deferred.promise;
+      });
     });
 
     it('will reject gets', () => {
-      return queryForRejection.get().then(
-        () => {
-          throw new Error('Promise resolved even though error was expected.');
-        },
-        err => {
-          expect(err.name).to.exist;
-          expect(err.message).to.exist;
-        }
-      );
-    });
-
-    it('will reject gets twice in a row', () => {
-      return queryForRejection
-        .get()
-        .then(
-          () => {
-            throw new Error('Promise resolved even though error was expected.');
-          },
-          err => {
-            expect(err.name).to.exist;
-            expect(err.message).to.exist;
-          }
-        )
-        .then(() => queryForRejection.get())
-        .then(
+      return withTestCollection(persistence, {}, queryForRejection => {
+        return queryForRejection.get().then(
           () => {
             throw new Error('Promise resolved even though error was expected.');
           },
@@ -887,6 +864,38 @@ apiDescribe('Database', (persistence: boolean) => {
             expect(err.message).to.exist;
           }
         );
+      });
+    });
+
+    it('will reject gets twice in a row', () => {
+      return withTestCollection(persistence, {}, queryForRejection => {
+        return queryForRejection
+          .get()
+          .then(
+            () => {
+              throw new Error(
+                'Promise resolved even though error was expected.'
+              );
+            },
+            err => {
+              expect(err.name).to.exist;
+              expect(err.message).to.exist;
+            }
+          )
+          .then(() => queryForRejection.get())
+          .then(
+            () => {
+              throw new Error(
+                'Promise resolved even though error was expected.'
+              );
+            },
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            (err: any) => {
+              expect(err.name).to.exist;
+              expect(err.message).to.exist;
+            }
+          );
+      });
     });
   });
 
@@ -1017,8 +1026,10 @@ apiDescribe('Database', (persistence: boolean) => {
         const options = app.options;
 
         await app.delete();
-        const app2 = firebase.initializeApp(options, name);
-        const firestore2 = firebase.firestore!(app2);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        const app2 = (firebase as any).initializeApp(options, name);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        const firestore2 = (firebase as any).firestore(app2);
         await firestore2.enablePersistence();
         const docRef2 = firestore2.doc(docRef.path);
         const docSnap2 = await docRef2.get({ source: 'cache' });
@@ -1040,35 +1051,15 @@ apiDescribe('Database', (persistence: boolean) => {
 
         await app.delete();
         await firestore.clearPersistence();
-        const app2 = firebase.initializeApp(options, name);
-        const firestore2 = firebase.firestore!(app2);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        const app2 = (firebase as any).initializeApp(options, name);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        const firestore2 = (firebase as any).firestore(app2);
         await firestore2.enablePersistence();
         const docRef2 = firestore2.doc(docRef.path);
         await expect(
           docRef2.get({ source: 'cache' })
         ).to.eventually.be.rejectedWith('Failed to get document from cache.');
-      });
-    }
-  );
-
-  // eslint-disable-next-line no-restricted-properties
-  (persistence ? it : it.skip)(
-    'will reject the promise if clear persistence fails',
-    async () => {
-      await withTestDoc(persistence, async docRef => {
-        const oldDelete = SimpleDb.delete;
-        try {
-          SimpleDb.delete = (name: string): Promise<void> => {
-            return Promise.reject('Failed to delete the database.');
-          };
-          const firestore = docRef.firestore;
-          await firestore.app.delete();
-          await expect(
-            firestore.clearPersistence()
-          ).to.eventually.be.rejectedWith('Failed to delete the database.');
-        } finally {
-          SimpleDb.delete = oldDelete;
-        }
       });
     }
   );
@@ -1121,7 +1112,8 @@ apiDescribe('Database', (persistence: boolean) => {
       const firestore = docRef.firestore;
       await firestore.terminate();
 
-      const newFirestore = firebase.firestore!(firestore.app);
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const newFirestore = (firebase as any).firestore(firestore.app);
       expect(newFirestore).to.not.equal(firestore);
 
       // New instance functions.
@@ -1197,25 +1189,6 @@ apiDescribe('Database', (persistence: boolean) => {
       await pendingWrites;
       await awaitPendingWrites;
     });
-  });
-
-  it('waiting for pending writes should fail when user changes', async () => {
-    await withMockCredentialProviderTestDb(
-      persistence,
-      async (db, mockCredentialsProvider) => {
-        // Prevent pending writes receiving acknowledgement.
-        await db.disableNetwork();
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises
-        db.doc('abc/123').set({ foo: 'bar' });
-        const awaitPendingWrite = db.waitForPendingWrites();
-
-        mockCredentialsProvider.triggerUserChange(new User('user_1'));
-
-        await expect(awaitPendingWrite).to.be.eventually.rejectedWith(
-          "'waitForPendingWrites' promise is rejected due to a user change."
-        );
-      }
-    );
   });
 
   it('waiting for pending writes resolves immediately when offline and no pending writes', async () => {
