+++ packages/firestore/test/integration/api/fields.test.ts
@@ -16,8 +16,7 @@
  */
 
 import { expect } from 'chai';
-import { LogLevel, getLogLevel, setLogLevel } from '../../../src/util/log';
-import firebase from '../util/firebase_export';
+import * as firebase from '../util/firebase_export';
 import {
   apiDescribe,
   DEFAULT_SETTINGS,
@@ -27,9 +26,12 @@ import {
   withTestDoc
 } from '../util/helpers';
 
-const FieldPath = firebase.firestore!.FieldPath;
-const FieldValue = firebase.firestore!.FieldValue;
-const Timestamp = firebase.firestore!.Timestamp;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const FieldPath = (firebase as any).firestore.FieldPath;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const FieldValue = (firebase as any).firestore.FieldValue;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const Timestamp = (firebase as any).firestore.Timestamp;
 
 // Allow custom types for testing.
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -354,18 +356,12 @@ apiDescribe('Timestamp Fields in snapshots', (persistence: boolean) => {
   };
 
   it('are returned as native dates if timestampsInSnapshots set to false', () => {
-    // Avoid the verbose log message triggered by timestampsInSnapshots ==
-    // false.
-    const logLevel = getLogLevel();
-    setLogLevel(LogLevel.SILENT);
-
     const settings = { ...DEFAULT_SETTINGS };
     settings['timestampsInSnapshots'] = false;
 
     const timestamp = new Timestamp(100, 123456789);
     const testDocs = { a: testDataWithTimestamps(timestamp) };
     return withTestCollectionSettings(persistence, settings, testDocs, coll => {
-      setLogLevel(logLevel);
       return coll
         .doc('a')
         .get()
@@ -420,9 +416,6 @@ apiDescribe('Timestamp Fields in snapshots', (persistence: boolean) => {
   });
 
   it('timestampsInSnapshots affects server timestamps', () => {
-    const logLevel = getLogLevel();
-    setLogLevel(LogLevel.SILENT);
-
     const settings = { ...DEFAULT_SETTINGS };
     settings['timestampsInSnapshots'] = false;
     const testDocs = {
@@ -430,7 +423,6 @@ apiDescribe('Timestamp Fields in snapshots', (persistence: boolean) => {
     };
 
     return withTestCollectionSettings(persistence, settings, testDocs, coll => {
-      setLogLevel(logLevel);
       return coll
         .doc('a')
         .get()
