+++ packages/firestore/test/integration/api/transactions.test.ts
@@ -17,11 +17,11 @@
 
 import * as firestore from '@firebase/firestore-types';
 import { expect } from 'chai';
-import { Deferred } from '../../util/promise';
-import firebase from '../util/firebase_export';
+import * as firebase from '../util/firebase_export';
 import * as integrationHelpers from '../util/helpers';
-import { asyncQueue } from '../util/internal_helpers';
-import { TimerId } from '../../../src/util/async_queue';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const FieldPath = (firebase as any).firestore.FieldPath;
 
 const apiDescribe = integrationHelpers.apiDescribe;
 apiDescribe('Database transactions', (persistence: boolean) => {
@@ -452,125 +452,6 @@ apiDescribe('Database transactions', (persistence: boolean) => {
     });
   });
 
-  it('increment transactionally', () => {
-    // A set of concurrent transactions.
-    const transactionPromises: Array<Promise<void>> = [];
-    const readPromises: Array<Promise<void>> = [];
-    // A barrier to make sure every transaction reaches the same spot.
-    const barrier = new Deferred();
-    let started = 0;
-
-    return integrationHelpers.withTestDb(persistence, db => {
-      asyncQueue(db).skipDelaysForTimerId(TimerId.RetryTransaction);
-      const doc = db.collection('counters').doc();
-      return doc
-        .set({
-          count: 5
-        })
-        .then(() => {
-          // Make 3 transactions that will all increment.
-          for (let i = 0; i < 3; i++) {
-            const resolveRead = new Deferred<void>();
-            readPromises.push(resolveRead.promise);
-            transactionPromises.push(
-              db.runTransaction(transaction => {
-                return transaction.get(doc).then(snapshot => {
-                  expect(snapshot).to.exist;
-                  started = started + 1;
-                  resolveRead.resolve();
-                  return barrier.promise.then(() => {
-                    transaction.set(doc, {
-                      count: snapshot.data()!['count'] + 1
-                    });
-                  });
-                });
-              })
-            );
-          }
-
-          // Let all of the transactions fetch the old value and stop once.
-          return Promise.all(readPromises);
-        })
-        .then(() => {
-          // Let all of the transactions continue and wait for them to
-          // finish.
-          expect(started).to.equal(3);
-          barrier.resolve();
-          return Promise.all(transactionPromises);
-        })
-        .then(() => {
-          // Now all transaction should be completed, so check the result.
-          return doc.get();
-        })
-        .then(snapshot => {
-          expect(snapshot).to.exist;
-          expect(snapshot.data()!['count']).to.equal(8);
-        });
-    });
-  });
-
-  it('update transactionally', () => {
-    // A set of concurrent transactions.
-    const transactionPromises: Array<Promise<void>> = [];
-    const readPromises: Array<Promise<void>> = [];
-    // A barrier to make sure every transaction reaches the same spot.
-    const barrier = new Deferred();
-    let counter = 0;
-
-    return integrationHelpers.withTestDb(persistence, db => {
-      asyncQueue(db).skipDelaysForTimerId(TimerId.RetryTransaction);
-      const doc = db.collection('counters').doc();
-      return doc
-        .set({
-          count: 5,
-          other: 'yes'
-        })
-        .then(() => {
-          // Make 3 transactions that will all increment.
-          for (let i = 0; i < 3; i++) {
-            const resolveRead = new Deferred<void>();
-            readPromises.push(resolveRead.promise);
-            transactionPromises.push(
-              db.runTransaction(transaction => {
-                return transaction.get(doc).then(snapshot => {
-                  expect(snapshot).to.exist;
-                  counter = counter + 1;
-                  resolveRead.resolve();
-                  return barrier.promise.then(() => {
-                    transaction.update(doc, {
-                      count: snapshot.data()!['count'] + 1
-                    });
-                  });
-                });
-              })
-            );
-          }
-
-          // Let all of the transactions fetch the old value and stop once.
-          return Promise.all(readPromises);
-        })
-        .then(() => {
-          // Let all of the transactions continue and wait for them to
-          // finish. There should be 3 initial transaction runs.
-          expect(counter).to.equal(3);
-          barrier.resolve();
-          return Promise.all(transactionPromises);
-        })
-        .then(() => {
-          // Now all transaction should be completed, so check the result.
-          // There should be a maximum of 3 retries: once for the 2nd update,
-          // and twice for the 3rd update.
-          expect(counter).to.be.lessThan(7);
-          return doc.get();
-        })
-        .then(snapshot => {
-          expect(snapshot).to.exist;
-          expect(snapshot.data()!['count']).to.equal(8);
-          expect(snapshot.data()!['other']).to.equal('yes');
-        });
-    });
-  });
-
   it('can update nested fields transactionally', () => {
     const initialData = {
       desc: 'Description',
@@ -592,7 +473,7 @@ apiDescribe('Database transactions', (persistence: boolean) => {
             doc,
             'owner.name',
             'Sebastian',
-            new firebase.firestore!.FieldPath('is.admin'),
+            new FieldPath('is.admin'),
             true
           );
         })
@@ -652,46 +533,6 @@ apiDescribe('Database transactions', (persistence: boolean) => {
             );
           })
       );
-    });
-  });
-
-  it('handle reading a doc twice with different versions', () => {
-    return integrationHelpers.withTestDb(persistence, db => {
-      asyncQueue(db).skipDelaysForTimerId(TimerId.RetryTransaction);
-      const doc = db.collection('counters').doc();
-      let counter = 0;
-      return doc
-        .set({
-          count: 15
-        })
-        .then(() => {
-          return db.runTransaction(transaction => {
-            counter++;
-            // Get the doc once.
-            return (
-              transaction
-                .get(doc)
-                // Do a write outside of the transaction. Because the transaction
-                // will retry, set the document to a different value each time.
-                .then(() => doc.set({ count: 1234 + counter }))
-                // Get the doc again in the transaction with the new
-                // version.
-                .then(() => transaction.get(doc))
-                // Now try to update the doc from within the transaction.
-                // This should fail, because we read 15 earlier.
-                .then(() => transaction.set(doc, { count: 16 }))
-            );
-          });
-        })
-        .then(() => expect.fail('transaction should fail'))
-        .catch(err => {
-          expect(err).to.exist;
-          expect((err as firestore.FirestoreError).code).to.equal('aborted');
-        })
-        .then(() => doc.get())
-        .then(snapshot => {
-          expect(snapshot.data()!['count']).to.equal(1234 + counter);
-        });
     });
   });
